chore: optimization
