+++ apps/frontend/src/app/2_molecules/AmountRenderer/AmountRenderer.utils.tsx
@@ -1,0 +1,18 @@
+import { Decimalish } from '@sovryn/utils';
+
+import { decimalic } from '../../../utils/math';
+
+export const isValueBetweenZeroAndOne = (value: number) =>
+  value > 0 && value < 1;
+
+export const calculateDecimalPlaces = (
+  value: Decimalish,
+  precision: number,
+) => {
+  const decimalValue = decimalic(value).toString();
+  const [, decimals = ''] = decimalValue.split('.');
+  const nonZeroIndex = decimals.search(/[^0]/);
+  return nonZeroIndex !== -1
+    ? Math.max(nonZeroIndex + 1, precision)
+    : precision;
+};
